backend/url_short_api/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
distlib==0.3.6
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.9.0
mypy==0.991
mypy-extensions==0.4.3
pipenv==2022.12.19
platformdirs==2.6.2
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
typing_extensions==4.4.0
virtualenv==20.17.1
virtualenv-clone==0.5.7
